+++ cli/src/commands/fix.rs
@@ -317,7 +317,11 @@ async fn fix_one_file(
         None
     } else {
         match &file_to_fix.base_file_id {
-            Some(base_file_id) if matching_tools.clone().any(|t| t.line_range_arg.is_some()) => {
+            Some(base_file_id)
+                if matching_tools
+                    .clone()
+                    .any(|t| !t.line_range_args.is_empty()) =>
+            {
                 let mut content = vec![];
                 let mut read = store
                     .read_file(&file_to_fix.repo_path, base_file_id)
@@ -332,7 +336,7 @@ async fn fix_one_file(
     let new_content = matching_tools.fold(old_content.clone(), |prev_content, tool_config| {
         let mut extra_args = Vec::new();
 
-        if let Some(line_range_arg) = &tool_config.line_range_arg {
+        if !tool_config.line_range_args.is_empty() {
             let RegionsToFormat::LineRanges(ranges) =
                 compute_regions_to_format(base_content.as_deref(), &prev_content);
             if ranges.is_empty() && !tool_config.run_tool_if_zero_line_ranges {
@@ -342,11 +346,13 @@ async fn fix_one_file(
             }
 
             for range in ranges {
-                extra_args.push(
-                    line_range_arg
-                        .replace("$first", &range.first.to_string())
-                        .replace("$last", &range.last.to_string()),
-                );
+                for line_range_arg in &tool_config.line_range_args {
+                    extra_args.push(
+                        line_range_arg
+                            .replace("$first", &range.first.to_string())
+                            .replace("$last", &range.last.to_string()),
+                    );
+                }
             }
         }
 
@@ -491,10 +497,11 @@ struct ToolConfig {
     /// Whether the tool is enabled
     enabled: bool,
     /// Arguments to pass changed line ranges to the tool.
-    /// The string is a template where `$first` and `$last` are replaced by the
-    /// 1-based first and last inclusive line numbers of the changed range.
-    /// For example, `"--lines=$first-$last"`.
-    line_range_arg: Option<String>,
+    /// Each string is a template where `$first` and `$last` are replaced by the
+    /// 1-based first and last inclusive line numbers of the changed range;
+    /// for example, `["--lines=$first:$last"]`. Each string will be passed as a
+    /// separate argument to the tool.
+    line_range_args: Vec<String>,
     /// Whether to run the tool invocation for this tool on files that had
     /// zero line ranges to format in the revision being fixed.
     ///
@@ -502,7 +509,7 @@ struct ToolConfig {
     /// This default behavior serves two main purposes:
     /// - Avoiding unnecessary executions of tools when no line ranges are
     ///   modified.
-    /// - Preventing tools configured with `line-range-arg` from making overly
+    /// - Preventing tools configured with `line-range-args` from making overly
     ///   broad changes (formatting the whole file) when no line ranges are
     ///   available.
     ///
@@ -530,7 +537,7 @@ struct RawToolConfig {
     #[serde(default = "default_tool_enabled")]
     enabled: bool,
     #[serde(default)]
-    line_range_arg: Option<String>,
+    line_range_args: Vec<String>,
     #[serde(default)]
     run_tool_if_zero_line_ranges: bool,
 }
@@ -562,9 +569,9 @@ fn get_tools_config(
                     .map(|arg| fileset::parse(&mut diagnostics, arg, fileset_context))
                     .try_collect()?,
             );
-            if tool.line_range_arg.is_none() && tool.run_tool_if_zero_line_ranges {
+            if tool.line_range_args.is_empty() && tool.run_tool_if_zero_line_ranges {
                 return Err(config_error(
-                    "run-tool-if-zero-line-ranges can only be set when line-range-arg is set",
+                    "run-tool-if-zero-line-ranges can only be set when line-range-args is set",
                 ));
             }
             print_parse_diagnostics(ui, &format!("In `fix.tools.{name}`"), &diagnostics)?;
@@ -572,7 +579,7 @@ fn get_tools_config(
                 command: tool.command,
                 matcher: expression.to_matcher(),
                 enabled: tool.enabled,
-                line_range_arg: tool.line_range_arg,
+                line_range_args: tool.line_range_args,
                 run_tool_if_zero_line_ranges: tool.run_tool_if_zero_line_ranges,
             })
         })