+++ cli/tests/test_fix_command.rs
@@ -1647,12 +1647,12 @@ fn test_fix_with_line_ranges() {
         [fix.tools.tool-1]
         command = [{formatter}, "--uppercase"]
         patterns = ["foo", "baz", "qux"]
-        line-range-arg = "--line-ranges=$first-$last"
+        line-range-args = ["--line-ranges=$first-$last"]
 
         [fix.tools.tool-2]
         command = [{formatter}, "--lowercase"]
         patterns = ["bar", "baz"]
-        line-range-arg = "--line-ranges=$first-$last"
+        line-range-args = ["--line-ranges", "$first-$last"]
         "###,
     ));
 
@@ -1667,7 +1667,7 @@ fn test_fix_with_line_ranges() {
     // Create new commit modifying foo and bar.
     work_dir.run_jj(["new"]).success();
     work_dir.write_file("foo", "Foo1\nFoo2-Mod\nFoo3\n");
-    work_dir.write_file("bar", "Bar1-Mod\nBar2\nBar3\n");
+    work_dir.write_file("bar", "Bar1-Mod\nBar2\nBar3-Mod\n");
     work_dir.write_file("baz", "unmodified\n");
     work_dir
         .run_jj(["bookmark", "create", "-r@", "c2"])
@@ -1688,8 +1688,8 @@ fn test_fix_with_line_ranges() {
     insta::assert_snapshot!(output, @"
     ------- stderr -------
     Fixed 2 commits of 2 checked.
-    Working copy  (@) now at: mzvwutvl 7d52a6e8 c3 | (no description set)
-    Parent commit (@-)      : kkmpptxz a8e4c2e8 c2 | (no description set)
+    Working copy  (@) now at: mzvwutvl 56f659c4 c3 | (no description set)
+    Parent commit (@-)      : kkmpptxz a3b54db4 c2 | (no description set)
     Added 0 files, modified 2 files, removed 0 files
     [EOF]
     ");
@@ -1727,7 +1727,7 @@ fn test_fix_with_line_ranges() {
     insta::assert_snapshot!(output, @"
     bar1-mod
     Bar2
-    Bar3
+    bar3-mod
     [EOF]
     ");
     let output = work_dir.run_jj(["file", "show", "baz", "-r", "c2"]);
@@ -1767,19 +1767,19 @@ fn test_fix_with_run_tool_if_zero_line_ranges() {
         [fix.tools.tool-1]
         command = [{formatter}, "--uppercase"]
         patterns = ["foo", "baz"]
-        line-range-arg = "--line-ranges=$first-$last"
+        line-range-args = ["--line-ranges=$first-$last"]
         run-tool-if-zero-line-ranges = true
 
         [fix.tools.tool-2]
         command = [{formatter}, "--lowercase"]
         patterns = ["bar"]
-        line-range-arg = "--line-ranges=$first-$last"
+        line-range-args = ["--line-ranges=$first-$last"]
         run-tool-if-zero-line-ranges = false
 
         [fix.tools.tool-3]
         command = [{formatter}, "--uppercase", "--split-even-length-lines"]
         patterns = ["qux"]
-        line-range-arg = "--line-ranges=$first-$last"
+        line-range-args = ["--line-ranges", "$first-$last"]
         run-tool-if-zero-line-ranges = true
         "###,
     ));
@@ -1899,6 +1899,49 @@ fn test_fix_with_run_tool_if_zero_line_ranges() {
 }
 
 #[test]
+fn test_fix_with_run_tool_if_zero_line_ranges_invalid() {
+    let test_env = TestEnvironment::default();
+    test_env.run_jj_in(".", ["git", "init", "repo"]).success();
+    let work_dir = test_env.work_dir("repo");
+    let formatter_path = assert_cmd::cargo::cargo_bin!("fake-formatter");
+    assert!(formatter_path.is_file());
+    let formatter = to_toml_value(formatter_path.to_str().unwrap());
+    test_env.add_config(format!(
+        r###"
+        [fix.tools.tool-1]
+        command = [{formatter}, "--uppercase"]
+        patterns = ["foo"]
+        line-range-args = []
+        run-tool-if-zero-line-ranges = true
+        "###,
+    ));
+
+    // Initial commit.
+    work_dir.write_file("foo", "Foo1\nFoo2\nFoo3\n");
+    work_dir
+        .run_jj(["bookmark", "create", "-r@", "c1"])
+        .success();
+
+    // Create a new commit modifying "foo".
+    work_dir.run_jj(["new"]).success();
+    work_dir.write_file("foo", "Foo1\nFoo3\n");
+    work_dir
+        .run_jj(["bookmark", "create", "-r@", "c2"])
+        .success();
+
+    // Run `jj fix` on the second commit. It should fail due to the invalid fix
+    // tools config.
+    let output = work_dir.run_jj(["fix", "-s", "c2"]);
+    insta::assert_snapshot!(output, @"
+    ------- stderr -------
+    Config error: run-tool-if-zero-line-ranges can only be set when line-range-args is set
+    For help, see https://docs.jj-vcs.dev/latest/config/ or use `jj help -k config`.
+    [EOF]
+    [exit status: 1]
+    ");
+}
+
+#[test]
 fn test_fix_with_all_lines_arg() {
     let test_env = TestEnvironment::default();
     test_env.run_jj_in(".", ["git", "init", "repo"]).success();
@@ -1911,7 +1954,7 @@ fn test_fix_with_all_lines_arg() {
         [fix.tools.tool-1]
         command = [{formatter}, "--uppercase"]
         patterns = ["foo", "baz", "qux"]
-        line-range-arg = "--line-ranges=$first-$last"
+        line-range-args = ["--line-ranges=$first-$last"]
 
         [fix.tools.tool-2]
         command = [{formatter}, "--lowercase"]
@@ -2013,12 +2056,12 @@ fn test_fix_with_line_ranges_multiple_formatters() {
         [fix.tools.tool-1]
         command = [{formatter}, "--split-even-length-lines"]
         patterns = ["foo", "boo"]
-        line-range-arg = "--line-ranges=$first-$last"
+        line-range-args = ["--line-ranges=$first-$last"]
 
         [fix.tools.tool-2]
         command = [{formatter}, "--uppercase"]
         patterns = ["foo", "boo"]
-        line-range-arg = "--line-ranges=$first-$last"
+        line-range-args = ["--line-ranges=$first-$last"]
         "###,
     ));
 
@@ -2089,7 +2132,7 @@ fn test_fix_with_line_ranges_and_include_unchanged_files_all_lines() {
         [fix.tools.tool-1]
         command = [{formatter}, "--uppercase"]
         patterns = ["all()"]
-        line-range-arg = "--line-ranges=$first-$last"
+        line-range-args = ["--line-ranges=$first-$last"]
         "###,
     ));
 
@@ -2174,4 +2217,36 @@ fn test_fix_with_line_ranges_and_include_unchanged_files_all_lines() {
     ");
     let output = work_dir.run_jj(["file", "show", "empty.txt", "-r", "c2"]);
     insta::assert_snapshot!(output, @"");
+}
+
+// TODO: Remove in jj 0.51+
+#[test]
+fn test_fix_line_range_args_migration() {
+    let test_env = TestEnvironment::default();
+    test_env.run_jj_in(".", ["git", "init", "repo"]).success();
+    let work_dir = test_env.work_dir("repo");
+    let formatter_path = assert_cmd::cargo::cargo_bin!("fake-formatter");
+    assert!(formatter_path.is_file());
+    let formatter = to_toml_value(formatter_path.to_str().unwrap());
+    test_env.add_config(format!(
+        r###"
+        [fix.tools.tool-1]
+        command = [{formatter}, "--uppercase"]
+        patterns = ["all()"]
+        line-range-arg = "--line-ranges=$first-$last"
+        "###,
+    ));
+
+    work_dir.write_file("file.txt", "foo\n");
+
+    let output = work_dir.run_jj(["fix"]).success();
+    insta::assert_snapshot!(output, @r#"
+    ------- stderr -------
+    Warning: Deprecated user-level config: fix.tools.tool-1.line-range-arg is updated to fix.tools.tool-1.line-range-args = ["--line-ranges=$first-$last"]
+    Fixed 1 commits of 1 checked.
+    Working copy  (@) now at: qpvuntsm bce2043c (no description set)
+    Parent commit (@-)      : zzzzzzzz 00000000 (empty) (no description set)
+    Added 0 files, modified 1 files, removed 0 files
+    [EOF]
+    "#);
 }